@@ -1,1 +1,5 @@
-<slot />
+<script>
+	let { children } = $props();
+</script>
+
+{@render children()}
